Add error handling

+++ src/author/author.service.ts
@@ -1,7 +1,9 @@
 import { PrismaService } from 'src/prisma/prisma.service';
 import { AddAuthorDto } from './dto';
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { EditAuthorDto } from './dto/edit-author.dto';
+import { NotFoundException } from 'src/exceptions/not-found.exceptions';
+import { InternalServerErrorException } from 'src/exceptions/internal-server-error.exception';
 
 @Injectable()
 export class AuthorService {
@@ -12,45 +14,79 @@ export class AuthorService {
   }
 
   async getAuthor(id: number) {
-    return await this.prisma.author.findUnique({
-      where: {
-        id,
-      },
-    });
-  }
-
-  async addAuthor(dto: AddAuthorDto) {
-    return await this.prisma.author.create({
-      data: {
-        name: dto.name,
-      },
-    });
-  }
-
-  async updateAuthor(id: number, dto: EditAuthorDto) {
     const author = await this.prisma.author.findUnique({
       where: {
         id,
       },
     });
+
     if (!author) {
-      throw new Error('Author not found');
+      throw new NotFoundException(
+        'Could not find an author with the provided id',
+      );
     }
-    return await this.prisma.author.update({
-      where: {
-        id,
-      },
-      data: {
-        name: dto.name || author.name,
-      },
-    });
+
+    return author;
+  }
+
+  async addAuthor(dto: AddAuthorDto) {
+    try {
+      return await this.prisma.author.create({
+        data: {
+          name: dto.name,
+        },
+      });
+    } catch (error) {
+      throw new InternalServerErrorException();
+    }
+  }
+
+  async updateAuthor(id: number, dto: EditAuthorDto) {
+    try {
+      const author = await this.prisma.author.findUnique({
+        where: {
+          id,
+        },
+      });
+      if (!author) {
+        throw new NotFoundException(
+          'Could not find an author with the provided id',
+        );
+      }
+      return await this.prisma.author.update({
+        where: {
+          id,
+        },
+        data: {
+          name: dto.name || author.name,
+        },
+      });
+    } catch (error) {
+      throw new InternalServerErrorException();
+    }
   }
 
   async deleteAuthor(id: number) {
-    return await this.prisma.author.delete({
+    const author = await this.prisma.author.findUnique({
       where: {
         id,
       },
     });
+
+    if (!author) {
+      throw new NotFoundException(
+        'There is no such an author in the database.',
+      );
+    }
+
+    try {
+      return await this.prisma.author.delete({
+        where: {
+          id,
+        },
+      });
+    } catch (error) {
+      throw new InternalServerErrorException();
+    }
   }
 }
